refactor(scripts): fold 5 findWorkspaceRoot copies into repo-root.mjs (#12333)
Add `findWorkspaceRoot(startDir)` to `packages/scripts/lib/repo-root.mjs` (+ its
`.d.ts`) — walk up until a package.json declares `workspaces`, falling back to
the resolved `startDir` at the filesystem root. Replace the five identical
copy-pasted helpers in ensure-tsc-nested-output-dir, flatten-tsc-package-output,
rewrite-dist-relative-imports-node-esm, verify-package-runtime-exports, and
with-package-build-lock with the shared import.

The consolidated helper is sync (each copy read a handful of tiny manifests);
the four async call sites drop their now-redundant `await`. All five callers
pass `process.cwd()`, for which the fallback is identical to the old
`process.cwd()` literal.

Behavior-preserving: each migrated script resolves the same workspace root and
produces identical output (verified against the pre-migration copies).
